requirements.txt
requests==2.32.*
psycopg2==2.9.*
atproto==0.0.*
pillow==11.2.*
